* libc/include/sys/unistd.h: Declare getpagesize as returning POSIX-mandated int. * libc/sys/linux/sys/unistd.h: Ditto.

+++ newlib/libc/include/sys/unistd.h
@@ -81,7 +81,7 @@ char    _EXFUN(*getlogin, (void ));
 int _EXFUN(getlogin_r, (char *name, size_t namesize) );
 #endif
 char 	_EXFUN(*getpass, (const char *__prompt));
-size_t  _EXFUN(getpagesize, (void));
+int	_EXFUN(getpagesize, (void));
 #if defined(__CYGWIN__)
 int    _EXFUN(getpeereid, (int, uid_t *, gid_t *));
 #endif
